requirements.txt
annotated-doc==0.0.4
click==8.3.1
colorama==0.4.6
markdown-it-py==4.0.0
mdurl==0.1.2
nanoid==2.0.0
psycopg[binary]==3.3.3
Pygments==2.19.2
python-dotenv==1.2.2
rich==14.3.3
shellingham==1.5.4
typer==0.24.0
tzdata==2025.3